relocate ROMS-specific adapters
